+++ test/integration/live_advanced_test.exs
@@ -30,8 +30,9 @@ defmodule DrabTestApp.LiveAdvancedTest do
       assert query_one!(fixture.socket, "#excluded", :innerText) == %{"innerText" => "Hegemon"}
     end
 
+    @tag capture_log: true
     test "update child should raise", fixture do
-      Code.compiler_options(warnings_as_errors: true)
+      # Code.compiler_options(warnings_as_errors: true)
       IO.puts("\n--> the following warning is expected:")
       assert_raise CompileError, fn -> poke(fixture.socket, excluded: "Hegemon") end
     end